MedToolAgent/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

langchain>=0.1.0
langchain-ollama>=0.1.0
langgraph>=0.0.20
mcp>=0.1.0

pydantic>=2.0.0
pydantic-settings>=2.0.0

python-dotenv>=1.0.0
requests>=2.31.0

biopython>=1.81
beautifulsoup4>=4.12.0
lxml>=4.9.0

faiss-cpu>=1.7.4
numpy>=1.24.0

google-generativeai>=0.3.0
streamlit>=1.28.0
